backend/sport_project/requirements.txt
django==3.2.23
djangorestframework==3.14.0
djongo==1.3.6
channels==4.0.0
channels-redis==4.1.0
sqlparse==0.2.4
six==1.16.0
pymongo==3.12.3